# docs/open-api-spec/swagger.yaml
definitions:
...
  controllers.ConsumerModel:
    properties:
      callbackURL:
        type: string
      changedAt:
        type: string
      deadLetterQueueURL:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
      type:
        type: string
    type: object
...
  controllers.ListResult:
    properties:
      links:
        additionalProperties:
          type: string
        type: object
      pages:
        additionalProperties:
          type: string
        type: object
      result:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: This is a fully HTTP-based Pub/Sub Broker designed to simplify system
    architecture in SOA or Microservice architecture.
  title: Webhook Broker API
  version: 1.0.0
paths:
...
  /channel/{channelId}/consumer/{consumerId}:
    delete:
      description: Deletes a consumer from a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Only delete if the resource has not been modified since this
          timestamp
        in: header
        name: If-Unmodified-Since
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "412":
          description: Precondition Failed
        "500":
          description: Internal Server Error
      summary: Delete Consumer
      tags:
      - Consumers
    get:
      description: Retrieves details for a specific consumer.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ConsumerModel'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Consumer Details
      tags:
      - Consumers
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Creates or updates a consumer for a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Only update if the resource has not been modified since this
          timestamp
        in: header
        name: If-Unmodified-Since
        type: string
      - description: Token
        in: formData
        name: token
        type: string
      - description: Name
        in: formData
        name: name
        type: string
      - description: Callback URL
        in: formData
        name: callbackUrl
        type: string
      - description: Consumer type
        in: formData
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ConsumerModel'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "412":
          description: Precondition Failed
        "415":
          description: Unsupported Media Type
        "500":
          description: Internal Server Error
      summary: Update Consumer
      tags:
      - Consumers
...
  /channel/{channelId}/consumers:
    get:
      description: Retrieves a paginated list of all consumers for a specific channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ListResult'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List Consumers for a Channel
      tags:
      - Consumers
...
securityDefinitions:
  ChannelToken:
    description: Channel authentication token issued when the channel was created
      or last rotated.
    in: header
    name: X-Broker-Channel-Token
    type: apiKey
  ConsumerToken:
    description: Consumer authentication token issued when the consumer was created
      or last rotated.
    in: header
    name: X-Broker-Consumer-Token
    type: apiKey
  ProducerID:
    description: Identifier of the producer originating the request.
    in: header
    name: X-Broker-Producer-ID
    type: apiKey
  ProducerToken:
    description: Producer authentication token issued when the producer was created
      or last rotated.
    in: header
    name: X-Broker-Producer-Token
    type: apiKey
swagger: "2.0"